skip failing tests on windows
