+++ nilearn/_utils/testing.py
@@ -9,6 +9,16 @@ import warnings
 from pathlib import Path
 
 import pytest
+from numpy import __version__ as np_version
+
+from nilearn._utils import compare_version
+from nilearn._utils.helpers import OPTIONAL_MATPLOTLIB_MIN_VERSION
+
+try:
+    from matplotlib import __version__ as mpl_version
+except ImportError:
+    mpl_version = OPTIONAL_MATPLOTLIB_MIN_VERSION
+
 
 # we use memory_profiler library for memory consumption checks
 try:
@@ -179,3 +189,11 @@ def skip_if_running_tests(msg=""):
     """
     if are_tests_running():
         pytest.skip(msg, allow_module_level=True)
+
+
+def on_windows_with_old_mpl_and_new_numpy():
+    return (
+        compare_version(np_version, ">", "1.26.4")
+        and compare_version(mpl_version, "<", "3.8.0")
+        and os.name == "nt"
+    )
