+++ nilearn/_utils/tests/test_helpers.py
@@ -15,6 +15,7 @@ from nilearn._utils.helpers import (
     rename_parameters,
     stringify_path,
 )
+from nilearn._utils.testing import on_windows_with_old_mpl_and_new_numpy
 
 
 def _mock_args_for_testing_replace_parameter():
@@ -35,6 +36,10 @@ def _mock_args_for_testing_replace_parameter():
 
 
 @pytest.mark.skipif(
+    on_windows_with_old_mpl_and_new_numpy(),
+    reason="Old matplotlib not compatible with numpy 2.0 on windows.",
+)
+@pytest.mark.skipif(
     is_matplotlib_installed(),
     reason="Test requires matplotlib not to be installed.",
 )
@@ -52,6 +57,10 @@ def test_should_raise_custom_warning_if_mpl_not_installed():
         _set_mpl_backend(warning)
 
 
+@pytest.mark.skipif(
+    on_windows_with_old_mpl_and_new_numpy(),
+    reason="Old matplotlib not compatible with numpy 2.0 on windows.",
+)
 @pytest.mark.skipif(
     is_matplotlib_installed(),
     reason="Test requires matplotlib not to be installed.",
