+++ nilearn/tests/test_package_import.py
@@ -1,27 +1,9 @@
-import os
 import sys
 
 import pytest
-from numpy import __version__ as np_version
 
-from nilearn._utils import compare_version
-from nilearn._utils.helpers import (
-    OPTIONAL_MATPLOTLIB_MIN_VERSION,
-    is_matplotlib_installed,
-)
-
-try:
-    from matplotlib import __version__ as mpl_version
-except ImportError:
-    mpl_version = OPTIONAL_MATPLOTLIB_MIN_VERSION
-
-
-def on_windows_with_old_mpl_and_new_numpy():
-    return (
-        compare_version(np_version, ">", "1.26.4")
-        and compare_version(mpl_version, "<", "3.8.0")
-        and os.name == "nt"
-    )
+from nilearn._utils.helpers import is_matplotlib_installed
+from nilearn._utils.testing import on_windows_with_old_mpl_and_new_numpy
 
 
 @pytest.mark.skipif(
